Matching Stockfish Beta having Timestamp: 1597415992 - Try to match relative magnitude of NNUE eval to classical.
Setting to use the Sergio's best network (20200720-1017.bin) as default. Notice that you will need this network to compile
Stockfish Polyglot but you won't find it in this GitHub repository because I'm not sure it's possible to reshare it.
You can download it at this address: https://www.comp.nus.edu.sg/~sergio-v/nnue/20200720-1017.bin .

Modifying the build.bat file to write the proper data suffix in the file name (DDMMYY).